+++ OUDS/Core/Themes/Orange/Sources/Values/OrangeBrandColorRawTokens+Values.swift
@@ -18,29 +18,29 @@ import OUDSTokensRaw
 
 extension OrangeBrandColorRawTokens {
 
-    // MARK: Primitive token - Colors - Orange - Brand - Orange
+    // MARK: Primitive token - Colors - Orange - Orange
 
-    public static let colorBrandOrange50: ColorOrangeBrandRawToken = "#FFF2E6"
-    public static let colorBrandOrange100: ColorOrangeBrandRawToken = "#FFD5B0"
-    public static let colorBrandOrange200: ColorOrangeBrandRawToken = "#FFC18A"
-    public static let colorBrandOrange300: ColorOrangeBrandRawToken = "#FFA554"
-    public static let colorBrandOrange400: ColorOrangeBrandRawToken = "#FF9433"
-    public static let colorBrandOrange500: ColorOrangeBrandRawToken = "#FF7900"
-    public static let colorBrandOrange550: ColorOrangeBrandRawToken = "#F16E00"
-    public static let colorBrandOrange600: ColorOrangeBrandRawToken = "#E86E00"
-    public static let colorBrandOrange700: ColorOrangeBrandRawToken = "#B55600"
-    public static let colorBrandOrange800: ColorOrangeBrandRawToken = "#8C4300"
-    public static let colorBrandOrange900: ColorOrangeBrandRawToken = "#6B3300"
+    public static let colorOrange50: ColorOrangeBrandRawToken = "#FFF2E6"
+    public static let colorOrange100: ColorOrangeBrandRawToken = "#FFD5B0"
+    public static let colorOrange200: ColorOrangeBrandRawToken = "#FFC18A"
+    public static let colorOrange300: ColorOrangeBrandRawToken = "#FFA554"
+    public static let colorOrange400: ColorOrangeBrandRawToken = "#FF9433"
+    public static let colorOrange500: ColorOrangeBrandRawToken = "#FF7900"
+    public static let colorOrange550: ColorOrangeBrandRawToken = "#F16E00"
+    public static let colorOrange600: ColorOrangeBrandRawToken = "#E86E00"
+    public static let colorOrange700: ColorOrangeBrandRawToken = "#B55600"
+    public static let colorOrange800: ColorOrangeBrandRawToken = "#8C4300"
+    public static let colorOrange900: ColorOrangeBrandRawToken = "#6B3300"
 
-    // MARK: Primitive token - Colors - Orange - Brand - Warm gray
+    // MARK: Primitive token - Colors - Orange - Warm gray
 
-    public static let colorBrandWarmGray100: ColorOrangeBrandRawToken = "#F9F5F0"
-    public static let colorBrandWarmGray200: ColorOrangeBrandRawToken = "#E9DDCE"
-    public static let colorBrandWarmGray300: ColorOrangeBrandRawToken = "#D6C4AE"
-    public static let colorBrandWarmGray400: ColorOrangeBrandRawToken = "#C1AB90"
-    public static let colorBrandWarmGray500: ColorOrangeBrandRawToken = "#A99275"
-    public static let colorBrandWarmGray600: ColorOrangeBrandRawToken = "#8A7860"
-    public static let colorBrandWarmGray700: ColorOrangeBrandRawToken = "#685D50"
-    public static let colorBrandWarmGray800: ColorOrangeBrandRawToken = "#48433D"
-    public static let colorBrandWarmGray900: ColorOrangeBrandRawToken = "#353228"
+    public static let colorWarmGray100: ColorOrangeBrandRawToken = "#F9F5F0"
+    public static let colorWarmGray200: ColorOrangeBrandRawToken = "#E9DDCE"
+    public static let colorWarmGray300: ColorOrangeBrandRawToken = "#D6C4AE"
+    public static let colorWarmGray400: ColorOrangeBrandRawToken = "#C1AB90"
+    public static let colorWarmGray500: ColorOrangeBrandRawToken = "#A99275"
+    public static let colorWarmGray600: ColorOrangeBrandRawToken = "#8A7860"
+    public static let colorWarmGray700: ColorOrangeBrandRawToken = "#685D50"
+    public static let colorWarmGray800: ColorOrangeBrandRawToken = "#48433D"
+    public static let colorWarmGray900: ColorOrangeBrandRawToken = "#353228"
 }